requirements.txt
APScheduler==3.7.0
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.2.1
sqlparse==0.2.4
tzlocal==2.1
urllib3==1.26.3
